+++ CHANGELOG.markdown
@@ -1,6 +1,8 @@
 2.2.2
 -----
 * `Control.Monad.Identity` now re-exports `Control.Monad.Trans.Identity`
+* Fix a bug in which `Control.Monad.State.Class.modify'` was not as strict in
+  the new state as its counterparts in `transformers`
 
 2.2.1
 -------